feat: added jarust context containing runtime
