--- src/raw_jacallback.rs
@@ -1,6 +1,0 @@
-use std::fmt::Debug;
-
-pub trait RawJaCallback: Send + Sync + Debug {
-    fn on_connection_success(&self);
-    fn on_connection_failure(&self);
-}
